+++ DependencyInjection/Configuration.php
@@ -20,7 +20,7 @@ use Symfony\Component\Config\Definition\ConfigurationInterface;
  *
  * @author Alexander <iam.asm89@gmail.com>
  */
-class Configuration implements ConfigurationInterface
+final class Configuration implements ConfigurationInterface
 {
     /**
      * Array of supported resource owners, indentation is intentional to easily notice
@@ -28,8 +28,8 @@ class Configuration implements ConfigurationInterface
      *
      * @var array
      */
-    private static $resourceOwners = array(
-        'oauth2' => array(
+    private static $resourceOwners = [
+        'oauth2' => [
             'amazon',
             'asana',
             'auth0',
@@ -48,6 +48,7 @@ class Configuration implements ConfigurationInterface
             'facebook',
             'fiware',
             'foursquare',
+            'genius',
             'github',
             'gitlab',
             'google',
@@ -81,8 +82,8 @@ class Configuration implements ConfigurationInterface
             'yandex',
             '37signals',
             'itembase',
-        ),
-        'oauth1' => array(
+        ],
+        'oauth1' => [
             'bitbucket',
             'discogs',
             'dropbox',
@@ -93,8 +94,8 @@ class Configuration implements ConfigurationInterface
             'twitter',
             'xing',
             'yahoo',
-        ),
-    );
+        ],
+    ];
 
     /**
      * Return the type (OAuth1 or OAuth2) of given resource owner.
@@ -110,7 +111,7 @@ class Configuration implements ConfigurationInterface
             return $resourceOwner;
         }
 
-        if (in_array($resourceOwner, static::$resourceOwners['oauth1'], true)) {
+        if (\in_array($resourceOwner, static::$resourceOwners['oauth1'], true)) {
             return 'oauth1';
         }
 
@@ -131,11 +132,11 @@ class Configuration implements ConfigurationInterface
             return true;
         }
 
-        if (in_array($resourceOwner, static::$resourceOwners['oauth1'], true)) {
+        if (\in_array($resourceOwner, static::$resourceOwners['oauth1'], true)) {
             return true;
         }
 
-        return in_array($resourceOwner, static::$resourceOwners['oauth2'], true);
+        return \in_array($resourceOwner, static::$resourceOwners['oauth2'], true);
     }
 
     /**
@@ -145,9 +146,15 @@ class Configuration implements ConfigurationInterface
      */
     public function getConfigTreeBuilder()
     {
-        $builder = new TreeBuilder();
+        $builder = new TreeBuilder('hwi_oauth');
 
-        $rootNode = $builder->root('hwi_oauth');
+        if (method_exists($builder, 'getRootNode')) {
+            $rootNode = $builder->getRootNode();
+        } else {
+            // BC layer for symfony/config 4.1 and older
+            $rootNode = $builder->root('hwi_oauth');
+        }
+
         $rootNode
             ->fixXmlConfig('firewall_name')
             ->children()
@@ -278,7 +285,7 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('type')
                                 ->validate()
                                     ->ifTrue(function ($type) {
-                                        return !Configuration::isResourceOwnerSupported($type);
+                                        return !self::isResourceOwnerSupported($type);
                                     })
                                     ->thenInvalid('Unknown resource owner type "%s".')
                                 ->end()
@@ -298,11 +305,11 @@ class Configuration implements ConfigurationInterface
                                                 return true;
                                             }
 
-                                            if (is_array($v)) {
-                                                return 0 === count($v);
+                                            if (\is_array($v)) {
+                                                return 0 === \count($v);
                                             }
 
-                                            if (is_string($v)) {
+                                            if (\is_string($v)) {
                                                 return empty($v);
                                             }
 
@@ -325,7 +332,7 @@ class Configuration implements ConfigurationInterface
                                 }
 
                                 // for each type at least these have to be set
-                                foreach (array('type', 'client_id', 'client_secret') as $child) {
+                                foreach (['type', 'client_id', 'client_secret'] as $child) {
                                     if (!isset($c[$child])) {
                                         return true;
                                     }
@@ -347,7 +354,7 @@ class Configuration implements ConfigurationInterface
                                     return false;
                                 }
 
-                                $children = array('authorization_url', 'access_token_url', 'request_token_url', 'infos_url');
+                                $children = ['authorization_url', 'access_token_url', 'request_token_url', 'infos_url'];
                                 foreach ($children as $child) {
                                     // This option exists only for OAuth1.0a
                                     if ('request_token_url' === $child && 'oauth2' === $c['type']) {
@@ -376,11 +383,11 @@ class Configuration implements ConfigurationInterface
                                 }
 
                                 // one of this two options must be set
-                                if (0 === count($c['paths'])) {
+                                if (0 === \count($c['paths'])) {
                                     return !isset($c['user_response_class']);
                                 }
 
-                                foreach (array('identifier', 'nickname', 'realname') as $child) {
+                                foreach (['identifier', 'nickname', 'realname'] as $child) {
                                     if (!isset($c['paths'][$child])) {
                                         return true;
                                     }
@@ -394,7 +401,7 @@ class Configuration implements ConfigurationInterface
                             ->ifTrue(function ($c) {
                                 if (isset($c['service'])) {
                                     // ignore paths & options if none were set
-                                    return 0 !== count($c['paths']) || 0 !== count($c['options']) || 3 < count($c);
+                                    return 0 !== \count($c['paths']) || 0 !== \count($c['options']) || 3 < \count($c);
                                 }
 
                                 return false;